requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.29.2
Django==5.0.4
django-multiupload==0.6.1
dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib_1697537062390/work
face-recognition==1.3.0
face_recognition_models==0.3.0
kombu==5.3.7
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225359967/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=031b7d6b2e5e604d9e21fc21be713ebf28ce133ec872dce6de006742d5e49bab
pillow==10.3.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
redis==5.0.4
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
